Remove unnecessary export of memset. NFC (#15617)
This code is kind of in the wrong place.  If we wanted to keep
it we could move it to emcc.py:2164 where we export symbols needed
by the sanitizers:

```
  if sanitize:
    settings.USE_OFFSET_CONVERTER = 1
    settings.EXPORTED_FUNCTIONS += [
        '_memalign',
        '_emscripten_builtin_memalign',
        '_emscripten_builtin_malloc',
        '_emscripten_builtin_free',
        '___heap_base',
        '___global_base'
    ]
```

However, I don't think we do need it.  The comment set that `memset` is
needed for `mmap` but `mmap` uses the `zeroMemory` JS function which
does not depend on `memset` (perhaps it once did?)