Avoid using assert except when ASSERTIONS is defined (#15588)
assert() is really designed to be used in ASSERTIONS mode
and should not be used in release.  This change updates all
the places in the codebase where we were doing this.

I'm not sure about that value of throwing `new Error` vs
just thrown raw strings in library_dylink.js but there is
already one place where new Error is used, and we can
switch them over later as needed.