Be consistent about how we add the emscripten root to `sys.path`. NFC (#15704)
Turns out it can be a little tricky to get this right so always use this
consistent pattern.

`toolchain_profiler.py` don't not need to do any sys.path manipulation
since it is not an entry point and should always be imported via the top
level import (i.e. `tools.toolchain_profiler`).