Error during `wasm-ld` when essential symbols are missing. NFC (#15716)
Create new  `REQUIRED_SYMBOLS` internal setting.  Any symbols in
this list are exported with `--export` rather than `--export-if-defined`
which means if they are missing the link will fail during `wasm-ld`.

Like `EXPORT_IF_DEFINED` this list contains native symbol names without
any `_` prefixes.

These are internal settings so this change should not be user-visible.
This allows for a simpler/shorter `wasm-ld` commands and earlier failure
when one of these required symbols in missing.

This is useful for #15703.