Remove unnecessary export of `emscripten_futex_wake`. NFC (#15767)
This was being exported solely for the benefit of an assertions that I
don't think served any purpose.

Its not clear exactly what this assertion was trying to check for but I
think it was checking for case where the Module object was not
initialized on the thread. However before this try/catch is even
executed we have already run several native methods. For example, see
`Module['__emscripten_thread_init']` just above.

For this reason I think removing the asserting and the extra export is
the correct action.