[NFC] Default system library building to -O2 (#15803)
This should avoid us ever forgetting to add opt flags to a system library
as was discovered in #15723. It also is a nice decrease in python code.

After that PR, all libraries have a specified opt level, and so this PR is
NFC: it just sets the default and removes code that specifies the same
value as the default.

As an extra check that this has no noticeable effect, I verified that code
size and metadce rebaselining with this patch leads to identical results
as without the patch. I also checked that the order of cflags is correct,
that is, when a library says -Oz that appears after the default -O2. And
a few tests break if that order were wrong (like test_metadce_cxx_except)
so we'd notice any mistakes here.