[SjLj] Make Wasm SjLj work with LTO (#15858)
This makes LTO work with Wasm SjLj.

For exceptions, we have `-fwasm-exceptions` option, which was created to
match other kinds of exceptions (DWARF, SjLj, and SEH). Clang driver
checks if `-fwasm-exceptions` is given, and if so does these things:
- Adds `target-features=+exception-handling` attribute to LLVM IR
- Adds `-mllvm -wasm-enable-eh` to the backend options.

For SjLj, we couldn't create such option like `-fwasm-exceptions`,
because other platforms all enable it by default and doesn't have a
dedicated Clang option for that. So we are passing
`-mllvm -wasm-enable-sjlj` in emcc to Clang as an workaround. Clang
driver checks if `-mllvm -wasm-enable-sjlj` is given, and if so adds
`target-features=+exception-handling` attribute to LLVM IR too. (Unlike
EH, it doesn't need to add `-mllvm -wasm-enable-sjlj` again because we
are already using that option)

The problem is, `-mllvm` options are not Clang options but LLVM options,
and we process them in `llvm_backend_args`. In non-LTO mode, these
options are added to Clang, but in LTO mode, these are added to wasm-ld.
But as mentioned above, we are using `-mllvm -wasm-enable-eh` as a kind
of Clang option due to lack of a dedicated Clang option like
`-fwasm-exceptions`. This adds `target-features=+exception-handling`
attribute to LLVM IR, which should be done in Clang and cannot be done
in wasm-ld.

We cannot just add `-mllvm -wasm-enable-eh` to `get_cflags`, because
then it will not be added to wasm-ld. The options is necessary in the
backend processing as well. We cannot add it to both `get_cflags` and
`llvm_backend_args` either because it will error out saying the same
option is given twice in non-LTO mode.

So this PR ads `-mexception-handling`, which enables EH feature, to
Clang, in case we use Wasm SjLj. (Wasm EH doesn't need this because
`-fwasm-exceptions` enables it in Clang.) This works for both LTO and
non-LTO modes.

Fixes #15665.